raising error in client and split as in routes_dp
