+++ client/pyproject.toml
@@ -27,6 +27,7 @@ disable = [
     "E0401",  # import-error
     "C0114",  # missing-module-docstring
     "W0102",  # dangerous-default-value
+    "R0801",  # duplicate-code
     "R0913",  # too-many-arguments
     "R0914"   # too-many-locals
 ]
